feat(email): add dkim signing capacity to message
a new module was created to provide dkim_signing capabilities to
message given a configuration. I tried to make as few changes as
possible in `Message` struct. It's rust beginner work, so comments are
more than welcome.
Using existing dkim create was on option but I did not find easy to
implement it in the existing code base. I started from the RFC and
implement it from scratch. Choices that were made to keep it manageable
to develop quickly.
* Only sha256 hashing with rsa signatures is implemented
* Repeating headers to be signed not supported since multiple headers
does not seem to be implemented in the lib
* l dkim paramtere was not implemented to keep things simple
* the sign function is not tested yet, didn't find a way to do so
properly using another implementation to check it