don't attempt to send commands when aborting
Currently abort() still tries to send a QUIT command on the smtp
connection before closing it. However, we may be in a state where
we can't write to the stream anymore and there is no other method
to close the connection.

For example, when upgrading the connection with TLS, we replace
the `inner` object with InnerAsyncNetworkStream::None. If the upgrade
fails, and if we try to abort the connection afterwards, we hit
a debug_assert() when writing.