feat(transport-smtp): add `message_iter` to `AsyncConnection` and `Connection`
The primary motivation for this change is memory usage - users of this crate
(like ourselves) might want to reduce memory usage while a email is
being sent. The current status in `lettre` is that you have to have the
whole message `&[u8]` in memory and does not allow any "lazier" approaches,
hence the new `message_iter` function.

A further improvement would be to make `ClientCodec` more transparent and
not require a buffer for its job, but I think this will be a huge
improvement for now.